Moved the Repository trait and the ManualRepository struct out of 'tuf' feature flag: 'tuf' lacks support for wasm due to its dependencies.
